Do not add SafeStackPass as a default CodeGen pass
SafeStack pass and adds it as a default CodeGen pass.  There is a
problem with it, though - a global variable
"__safestack_unsafe_stack_ptr" is declared by default, but is defined
only when SafeStack instrumentation is added for some function.  This
throws an undefined reference error almost always.

Disabling this pass for the current rebase.  It can be added back once
r240321 is added back (b/22349095)

Change-Id: I1857f4e6dcc52ca7aa5254134528e6443e04ca64